feat(ui,map): add signature view - including manually adding signatures, pasting signatures and custom controls necessary

+++ protocol/shared/src/main/scala/org/updraft0/controltower/protocol/system.scala
@@ -20,7 +20,7 @@ case class Station(
     corporationName: String
 )
 
-case class WormholeStatic(typeId: Long, name: String)
+case class WormholeStatic(typeId: Int, name: String)
 
 case class SolarSystem(
     id: Long,
